move sql/compiler optimizer files
