+++ db-engine/src/main/java/com/splicemachine/db/impl/sql/compile/optimizer/Level2CostEstimateImpl.java
@@ -29,9 +29,10 @@
  * and are licensed to you under the GNU Affero General Public License.
  */
 
-package com.splicemachine.db.impl.sql.compile;
+package com.splicemachine.db.impl.sql.compile.optimizer;
 
 import com.splicemachine.db.iapi.sql.compile.CostEstimate;
+import com.splicemachine.db.impl.sql.compile.CostEstimateImpl;
 
 public class Level2CostEstimateImpl extends CostEstimateImpl {
     public Level2CostEstimateImpl() { }
